+++ app/components/image-view-card.vue
@@ -1,42 +1,44 @@
 <template>
-  <div class="mb-2 mx-2 flex flex-col group">
-    <div class="relative">
-      <img
-        :src="encodedUrl"
-        loading="lazy"
-        class="w-full aspect-[334/187.88] rounded-lg mb-2 outline outline-1 outline-[--outline] group-hover:opacity-50 transition-opacity duration-300"
-      >
-      <div
-        class="absolute top-[10px] right-[10px] inset-0 flex gap-[10px] items-center justify-center opacity-0 group-hover:opacity-100 transition-opacity duration-300"
-      >
-        <card-preview-button
-          @preview="openPreview"
-        />
-        <card-download-button
-          :id="actionId"
-          :alt="props.alt"
-          :url="props.url"
-        />
-        <card-copy-button
-          :id="actionId"
-          :alt="props.alt"
-          :url="props.url"
-        />
+  <div>
+    <div class="mb-2 mx-2 flex flex-col group">
+      <div class="relative">
+        <img
+          :src="encodedUrl"
+          loading="lazy"
+          class="w-full aspect-[334/187.88] rounded-lg mb-2 outline outline-1 outline-[--outline] group-hover:opacity-50 transition-opacity duration-300"
+        >
+        <div
+          class="absolute top-[10px] right-[10px] inset-0 flex gap-[10px] items-center justify-center opacity-0 group-hover:opacity-100 transition-opacity duration-300"
+        >
+          <card-preview-button
+            @preview="openPreview"
+          />
+          <card-download-button
+            :id="actionId"
+            :alt="props.alt"
+            :url="props.url"
+          />
+          <card-copy-button
+            :id="actionId"
+            :alt="props.alt"
+            :url="props.url"
+          />
+        </div>
       </div>
+      <card-alt-description
+        :alt="props.alt"
+        :url="props.url"
+        :matches="props.matches"
+      />
     </div>
-    <card-alt-description
+    <card-preview-panel
+      v-model="isPreviewOpen"
       :alt="props.alt"
-      :url="props.url"
-      :matches="props.matches"
-    />
+      :image-url="encodedUrl"
+    >
+      {{ previewDescription }}
+    </card-preview-panel>
   </div>
-  <card-preview-panel
-    v-model="isPreviewOpen"
-    :alt="props.alt"
-    :image-url="encodedUrl"
-  >
-    {{ previewDescription }}
-  </card-preview-panel>
 </template>
 
 <script setup lang="ts">